Extract a dictionary from a data file with dic information

Extract a dictionary from a data file with dic information

Usage

extract_dic(data)

Arguments

data

A data frame with dic information.

Value

A data frame in a dictionary format.

Examples

extract_dic(ex_itrf)
#>    item_name
#> 1   itrf_I_1
#> 2   itrf_I_2
#> 3   itrf_I_4
#> 4   itrf_I_5
#> 5   itrf_I_6
#> 6   itrf_I_7
#> 7   itrf_I_8
#> 8   itrf_I_9
#> 9  itrf_I_10
#> 10 itrf_I_11
#> 11 itrf_I_12
#> 12 itrf_I_13
#> 13 itrf_I_14
#> 14 itrf_I_15
#> 15 itrf_I_16
#> 16 itrf_I_17
#> 17 itrf_I_19
#> 18 itrf_I_20
#> 19 itrf_I_23
#> 20 itrf_I_24
#> 21  itrf_E_1
#> 22  itrf_E_2
#> 23  itrf_E_3
#> 24  itrf_E_4
#> 25  itrf_E_5
#> 26  itrf_E_6
#> 27  itrf_E_7
#> 28  itrf_E_8
#> 29  itrf_E_9
#> 30 itrf_E_10
#> 31 itrf_E_11
#> 32 itrf_E_12
#> 33 itrf_E_13
#> 34 itrf_E_14
#> 35 itrf_E_15
#> 36 itrf_E_16
#> 39  itrf_int
#> 40  itrf_ext
#>                                                                                              item_label
#> 1                                                                        Verbringt zu viel Zeit alleine
#> 2                                                          Beschwert sich über Krankheit oder Schmerzen
#> 3                                                                       Vermeidet soziale Interaktionen
#> 4                                                                              Spielt bevorzugt alleine
#> 5                                 Geht nicht auf Kontaktversuche der Mitschülerinnen und Mitschüler ein
#> 6                                                             Macht sich Sorgen über unwichtige Details
#> 7                                                 Beschwert sich über Kopfschmerzen oder Bauchschmerzen
#> 8                                                                        Wirkt unglücklich oder traurig
#> 9                                                                           Klammert sich an Erwachsene
#> 10                                                                                  Verhält sich nervös
#> 11                                                                               Verhält sich ängstlich
#> 12                                                               Behauptet sich nicht gegenüber anderen
#> 13                                                                    Verhält sich übermäßig schüchtern
#> 14                                                                            Beklagt sich oder jammert
#> 15                                                              Beteiligt sich nicht an Gruppenaktionen
#> 16                                                                           Macht sich selbst schlecht
#> 17                                                                            Weint oder ist weinerlich
#> 18                                                                                Ist schnell verärgert
#> 19                                                                            Macht sich ständig Sorgen
#> 20                                                             Lässt sich langsam auf neue Personen ein
#> 21                                                  Stellt Unterrichtsaufgaben nicht rechtzeitig fertig
#> 22                                              Beginnt mit der Aufgabenbearbeitung nicht selbstständig
#> 23                                                                  Erledigt Hausaufgaben unvollständig
#> 24                                                    Zeigt Unterrichtsaufgaben nicht selbstständig vor
#> 25                                                               Kontrolliert seine eigene Arbeit nicht
#> 26 Nimmt Materialien, die zu Hause benötigt werden, nicht mit (z.B. für die Hausaufgaben, Elternbriefe)
#> 27                                                                   Streitet und zankt mit Lehrkräften
#> 28                                                                            Verliert die Beherrschung
#> 29                                                                                         Stört andere
#> 30                                                                      Verwendet unangemessene Sprache
#> 31                                                    Hat Konflikte mit Mitschülerinnen und Mitschülern
#> 32                                                                                      Kommandiert rum
#> 33                                                                      Macht unangebrachte Bemerkungen
#> 34                                                                   Kommt unvorbereitet zum Unterricht
#> 35                                                                   Beteiligt sich nicht am Unterricht
#> 36                                                           Respektiert nicht die Privatsphäre anderer
#> 39                                                                                   ITRF Internalizing
#> 40                                                                                   ITRF Externalizing
#>    values
#> 1     0:3
#> 2     0:3
#> 3     0:3
#> 4     0:3
#> 5     0:3
#> 6     0:3
#> 7     0:3
#> 8     0:3
#> 9     0:3
#> 10    0:3
#> 11    0:3
#> 12    0:3
#> 13    0:3
#> 14    0:3
#> 15    0:3
#> 16    0:3
#> 17    0:3
#> 18    0:3
#> 19    0:3
#> 20    0:3
#> 21    0:3
#> 22    0:3
#> 23    0:3
#> 24    0:3
#> 25    0:3
#> 26    0:3
#> 27    0:3
#> 28    0:3
#> 29    0:3
#> 30    0:3
#> 31    0:3
#> 32    0:3
#> 33    0:3
#> 34    0:3
#> 35    0:3
#> 36    0:3
#> 39   <NA>
#> 40   <NA>
#>                                                                                value_labels
#> 1  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 2  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 3  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 4  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 5  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 6  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 7  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 8  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 9  0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 10 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 11 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 12 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 13 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 14 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 15 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 16 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 17 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 18 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 19 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 20 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 21 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 22 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 23 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 24 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 25 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 26 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 27 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 28 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 29 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 30 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 31 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 32 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 33 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 34 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 35 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 36 0 = not problematic; 1 = slightly problematic; 2 = problematic; 3 = strongly problematic
#> 39                                                                                     <NA>
#> 40                                                                                     <NA>
#>    missing weight scale subscale subscale_2                    scale_label
#> 1      -99      1  ITRF      Int         SW Integrated teacher report form
#> 2      -99      1  ITRF      Int         AD Integrated teacher report form
#> 3      -99      1  ITRF      Int         SW Integrated teacher report form
#> 4      -99      1  ITRF      Int         SW Integrated teacher report form
#> 5      -99      1  ITRF      Int         SW Integrated teacher report form
#> 6      -99      1  ITRF      Int         AD Integrated teacher report form
#> 7      -99      1  ITRF      Int         AD Integrated teacher report form
#> 8      -99      1  ITRF      Int         AD Integrated teacher report form
#> 9      -99      1  ITRF      Int         AD Integrated teacher report form
#> 10     -99      1  ITRF      Int         AD Integrated teacher report form
#> 11     -99      1  ITRF      Int         AD Integrated teacher report form
#> 12     -99      1  ITRF      Int         SW Integrated teacher report form
#> 13     -99      1  ITRF      Int         SW Integrated teacher report form
#> 14     -99      1  ITRF      Int         AD Integrated teacher report form
#> 15     -99      1  ITRF      Int         SW Integrated teacher report form
#> 16     -99      1  ITRF      Int         AD Integrated teacher report form
#> 17     -99      1  ITRF      Int         AD Integrated teacher report form
#> 18     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 19     -99      1  ITRF      Int         AD Integrated teacher report form
#> 20     -99      1  ITRF      Int         SW Integrated teacher report form
#> 21     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 22     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 23     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 24     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 25     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 26     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 27     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 28     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 29     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 30     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 31     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 32     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 33     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 34     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 35     -99      1  ITRF      Ext        APD Integrated teacher report form
#> 36     -99      1  ITRF      Ext        OPP Integrated teacher report form
#> 39    <NA>     NA  ITRF      Int       <NA> Integrated teacher report form
#> 40    <NA>     NA  ITRF      Ext       <NA> Integrated teacher report form
#>    subscale_label                      subscale_2_label    type
#> 1   Internalizing                    Socially Withdrawn integer
#> 2   Internalizing                     Anxious/Depressed integer
#> 3   Internalizing                    Socially Withdrawn integer
#> 4   Internalizing                    Socially Withdrawn integer
#> 5   Internalizing                    Socially Withdrawn integer
#> 6   Internalizing                     Anxious/Depressed integer
#> 7   Internalizing                     Anxious/Depressed integer
#> 8   Internalizing                     Anxious/Depressed integer
#> 9   Internalizing                     Anxious/Depressed integer
#> 10  Internalizing                     Anxious/Depressed integer
#> 11  Internalizing                     Anxious/Depressed integer
#> 12  Internalizing                    Socially Withdrawn integer
#> 13  Internalizing                    Socially Withdrawn integer
#> 14  Internalizing                     Anxious/Depressed integer
#> 15  Internalizing                    Socially Withdrawn integer
#> 16  Internalizing                     Anxious/Depressed integer
#> 17  Internalizing                     Anxious/Depressed integer
#> 18  Externalizing               Oppositional/Disruptive integer
#> 19  Internalizing                     Anxious/Depressed integer
#> 20  Internalizing                    Socially Withdrawn integer
#> 21  Externalizing Academic Productivity/Disorganization integer
#> 22  Externalizing Academic Productivity/Disorganization integer
#> 23  Externalizing Academic Productivity/Disorganization integer
#> 24  Externalizing Academic Productivity/Disorganization integer
#> 25  Externalizing Academic Productivity/Disorganization integer
#> 26  Externalizing Academic Productivity/Disorganization integer
#> 27  Externalizing               Oppositional/Disruptive integer
#> 28  Externalizing               Oppositional/Disruptive integer
#> 29  Externalizing               Oppositional/Disruptive integer
#> 30  Externalizing               Oppositional/Disruptive integer
#> 31  Externalizing               Oppositional/Disruptive integer
#> 32  Externalizing               Oppositional/Disruptive integer
#> 33  Externalizing               Oppositional/Disruptive integer
#> 34  Externalizing Academic Productivity/Disorganization integer
#> 35  Externalizing Academic Productivity/Disorganization integer
#> 36  Externalizing               Oppositional/Disruptive integer
#> 39  Internalizing                                  <NA> integer
#> 40  Externalizing                                  <NA> integer
#>                               source class                        score_filter
#> 1  Volpe et al. et al. (unpublished)  item                                <NA>
#> 2  Volpe et al. et al. (unpublished)  item                                <NA>
#> 3  Volpe et al. et al. (unpublished)  item                                <NA>
#> 4  Volpe et al. et al. (unpublished)  item                                <NA>
#> 5  Volpe et al. et al. (unpublished)  item                                <NA>
#> 6  Volpe et al. et al. (unpublished)  item                                <NA>
#> 7  Volpe et al. et al. (unpublished)  item                                <NA>
#> 8  Volpe et al. et al. (unpublished)  item                                <NA>
#> 9  Volpe et al. et al. (unpublished)  item                                <NA>
#> 10 Volpe et al. et al. (unpublished)  item                                <NA>
#> 11 Volpe et al. et al. (unpublished)  item                                <NA>
#> 12 Volpe et al. et al. (unpublished)  item                                <NA>
#> 13 Volpe et al. et al. (unpublished)  item                                <NA>
#> 14 Volpe et al. et al. (unpublished)  item                                <NA>
#> 15 Volpe et al. et al. (unpublished)  item                                <NA>
#> 16 Volpe et al. et al. (unpublished)  item                                <NA>
#> 17 Volpe et al. et al. (unpublished)  item                                <NA>
#> 18 Volpe et al. et al. (unpublished)  item                                <NA>
#> 19 Volpe et al. et al. (unpublished)  item                                <NA>
#> 20 Volpe et al. et al. (unpublished)  item                                <NA>
#> 21 Volpe et al. et al. (unpublished)  item                                <NA>
#> 22 Volpe et al. et al. (unpublished)  item                                <NA>
#> 23 Volpe et al. et al. (unpublished)  item                                <NA>
#> 24 Volpe et al. et al. (unpublished)  item                                <NA>
#> 25 Volpe et al. et al. (unpublished)  item                                <NA>
#> 26 Volpe et al. et al. (unpublished)  item                                <NA>
#> 27 Volpe et al. et al. (unpublished)  item                                <NA>
#> 28 Volpe et al. et al. (unpublished)  item                                <NA>
#> 29 Volpe et al. et al. (unpublished)  item                                <NA>
#> 30 Volpe et al. et al. (unpublished)  item                                <NA>
#> 31 Volpe et al. et al. (unpublished)  item                                <NA>
#> 32 Volpe et al. et al. (unpublished)  item                                <NA>
#> 33 Volpe et al. et al. (unpublished)  item                                <NA>
#> 34 Volpe et al. et al. (unpublished)  item                                <NA>
#> 35 Volpe et al. et al. (unpublished)  item                                <NA>
#> 36 Volpe et al. et al. (unpublished)  item                                <NA>
#> 39                              <NA> scale scale == "ITRF" & subscale == "Int"
#> 40                              <NA> scale scale == "ITRF" & subscale == "Ext"
#>    score_function index         note
#> 1            <NA>     1         <NA>
#> 2            <NA>     2         <NA>
#> 3            <NA>     4         <NA>
#> 4            <NA>     5         <NA>
#> 5            <NA>     6         <NA>
#> 6            <NA>     7         <NA>
#> 7            <NA>     8         <NA>
#> 8            <NA>     9         <NA>
#> 9            <NA>    10         <NA>
#> 10           <NA>    11         <NA>
#> 11           <NA>    12         <NA>
#> 12           <NA>    13         <NA>
#> 13           <NA>    14         <NA>
#> 14           <NA>    15         <NA>
#> 15           <NA>    16         <NA>
#> 16           <NA>    17         <NA>
#> 17           <NA>    19         <NA>
#> 18           <NA>    17         <NA>
#> 19           <NA>    23         <NA>
#> 20           <NA>    24         <NA>
#> 21           <NA>     1         <NA>
#> 22           <NA>     2         <NA>
#> 23           <NA>     3         <NA>
#> 24           <NA>     4         <NA>
#> 25           <NA>     5         <NA>
#> 26           <NA>     6         <NA>
#> 27           <NA>     7         <NA>
#> 28           <NA>     8         <NA>
#> 29           <NA>     9         <NA>
#> 30           <NA>    10         <NA>
#> 31           <NA>    11         <NA>
#> 32           <NA>    12         <NA>
#> 33           <NA>    13         <NA>
#> 34           <NA>    14         <NA>
#> 35           <NA>    15 Item doubled
#> 36           <NA>    16         <NA>
#> 39           mean    NA         <NA>
#> 40           mean    NA         <NA>