Extracts a list of item names based on dic information

Extracts a list of item names based on dic information

get_scales(data, ..., .variable = "item_name")

Arguments

data

A data.frame with dic information.

...

selection definitions.

.variable

character with name of the dic information to be returned. Default is "item_name".

Value

A (named) list with string vectors of item names

Details

This function is basically a wrapper around the ` select_items(data = data, filter = ..., names_only = TRUE)` function. It takes multiple filter expressions for a single data frame and returns the item names. This is mostly helpful for functions that take multiple scale definitions like `alpha_table()`

Examples

get_scales(ex_itrf,
  'APD' = subscale_2 == "APD",
  'OPP' = subscale_2 == "OPP",
  "SW" = subscale_2 == "SW",
  'AD' = subscale_2 == "AD"
)
#> $APD
#>    itrf_E_1    itrf_E_2    itrf_E_3    itrf_E_4    itrf_E_5    itrf_E_6 
#>  "itrf_E_1"  "itrf_E_2"  "itrf_E_3"  "itrf_E_4"  "itrf_E_5"  "itrf_E_6" 
#>   itrf_E_14   itrf_E_15 
#> "itrf_E_14" "itrf_E_15" 
#> 
#> $OPP
#>   itrf_I_20    itrf_E_7    itrf_E_8    itrf_E_9   itrf_E_10   itrf_E_11 
#> "itrf_I_20"  "itrf_E_7"  "itrf_E_8"  "itrf_E_9" "itrf_E_10" "itrf_E_11" 
#>   itrf_E_12   itrf_E_13   itrf_E_16 
#> "itrf_E_12" "itrf_E_13" "itrf_E_16" 
#> 
#> $SW
#>    itrf_I_1    itrf_I_4    itrf_I_5    itrf_I_6   itrf_I_13   itrf_I_14 
#>  "itrf_I_1"  "itrf_I_4"  "itrf_I_5"  "itrf_I_6" "itrf_I_13" "itrf_I_14" 
#>   itrf_I_16   itrf_I_24 
#> "itrf_I_16" "itrf_I_24" 
#> 
#> $AD
#>    itrf_I_2    itrf_I_7    itrf_I_8    itrf_I_9   itrf_I_10   itrf_I_11 
#>  "itrf_I_2"  "itrf_I_7"  "itrf_I_8"  "itrf_I_9" "itrf_I_10" "itrf_I_11" 
#>   itrf_I_12   itrf_I_15   itrf_I_17   itrf_I_19   itrf_I_23 
#> "itrf_I_12" "itrf_I_15" "itrf_I_17" "itrf_I_19" "itrf_I_23" 
#>