Lookup normtable values

Transforms raw values to norm values based on a normtable.

lookup_norms(
  rawscores,
  group,
  normtable,
  from = "raw",
  to = "T",
  group_label = "group"
)

Arguments

rawscores

A vector with rawscores.

group

A vector with group affiliation.

normtable

An excel file name or a dataframe containing a normtable.

from

Label of the raw score variable in file.

to

Label of the norm score variable in file.

group_label

Label of the group variable in file.

Value

A vector with norm values

Examples

normtable <- data.frame(
  group = c(6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8),
  raw = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
  T = c(40, 42, 44, 46, 48, 50, 52, 54, 56, 58, 60, 37, 39, 41, 43, 45, 47,
        49, 51, 53, 55, 57)
)
rawscores <- c(5,5,3,1)
group <- c("6", "8", "6", "8")
lookup_norms(rawscores, group, normtable)
#> [1] 50 47 46 39

## When no group values are provided, raw scores are ambiguous:
lookup_norms(rawscores, normtable = normtable)
#> Warning: Multiple values found for raw 5. NA returned.
#> Warning: Multiple values found for raw 5. NA returned.
#> Warning: Multiple values found for raw 3. NA returned.
#> Warning: Multiple values found for raw 1. NA returned.
#> [1] NA NA NA NA