Table with alpha values

Returns a data.frame with item analyses for the provided scales.

alpha_table(
  data,
  scales,
  labels = NULL,
  round = 2,
  CI = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = "auto",
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE
)

Arguments

data

A data Frame

scales

A list containing vectors with variable names. Each list element defines one scale. Named list elements are used as labels.

labels

Label names for scales (defaults to named list elements in 'scales').

round

Rounds values to given decimal position.

CI

If TRUE confidence intervals are calculated.

conf_level

Confidence level (e.g. 0.95 for 95 percent).

check_key

Check_key for the psych::alpha function.

keys

Optional key argument for the psych::alpha function.

RMSEA

If TRUE RMSEA is calculated.

difficulty

If TRUE, the difficulty of the item is calculated.

values

Sets maximum and minimum valid values necessary to calculate item difficulty

fa

If TRUE, a one factor exploratory factor analyses is calculated and loadings are reported.

Value

A data frame with concise scale indices.

Examples

scales <- get_scales(ex_itrf,
  Int = scale == "ITRF" & subscale == "Int",
  Ext = scale == "ITRF" & subscale == "Ext"
)
alpha_table(ex_itrf, scales = scales, difficulty = TRUE, values = list(c(0, 3)), RMSEA = TRUE)
#> Note. values in brackets depict upper and lower bound of confidence intervals or [min,max] intervals.
#>   Scale    n n items    Alpha CI95% Std.Alph CI95% Homogeneity Discriminations
#> 1   Int 4247      19 .91 [.91, .92] .91 [.91, .92]         .36      [.49, .68]
#> 2   Ext 4616      17 .93 [.93, .93] .93 [.93, .93]         .44      [.39, .74]
#>   Difficulties        Means          SDs |Loadings| RMSEA
#> 1   [.07, .16] [0.21, 0.48] [0.51, 0.77] [.52, .70]  .154
#> 2   [.12, .32] [0.35, 0.95] [0.72, 1.04] [.38, .78]  .196