Scale Dictionary

A collection of procedures to bring labels, scales, missing values etc. to a data frame.

Author

Juergen Wilbert [aut, cre]

Examples

# apply a dictionary file to a data frame
dat <- apply_dic(dat_itrf, dic_itrf)

#> 2 missing types found and replaced with 'integer'.

#> Replaced missing values.

#> Values checked.

#> Scales scored.

#> `label` attribute set.

# check for typos (not allowed values)
dat <- check_values(dat, replace = NA)
# Single imputation (EM algorith from the Amelia package)
# based on the variables of the provided scale
dat <- impute_missing(dat, scale == "ITRF" & subscale == "Ext")

#> -- Imputation 1 --
#> 
#>   1  2  3
#> 

dat <- impute_missing(dat, scale == "ITRF" & subscale == "Int")

#> Warning: There are observations in the data that are completely missing. 
#>          These observations will remain unimputed in the final datasets. 
#> -- Imputation 1 --
#> 
#>   1  2  3
#> 

# Show a table with all scales and scale labels included in the data frame
list_scales(dat, labels = TRUE)

#> # A tibble: 4 × 6
#>   scale subscale subscale_2 scale_label      subscale_label subscale_2_label    
#>   <chr> <chr>    <chr>      <chr>            <chr>          <chr>               
#> 1 ITRF  Int      SW         Integrated teac… Internalizing  Socially Withdrawn  
#> 2 ITRF  Int      AD         Integrated teac… Internalizing  Anxious/Depressed   
#> 3 ITRF  Ext      OPP        Integrated teac… Externalizing  Oppositional/Disrup…
#> 4 ITRF  Ext      APD        Integrated teac… Externalizing  Academic Productivi…

# Example with pipeline syntax. Would be much easier to use the "describe" function
# from the psch packages instead of summarise_all here.
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dat %>%
  select_items(scale == "ITRF" & subscale == "Ext") %>%
  rename_items(pattern = c("name", "subscale_2")) %>%
  summarise_all(mean, na.rm = TRUE) %>%
  round(2) %>%
  t()

#>               [,1]
#> itrf_I_20_OPP 0.55
#> itrf_E_1_APD  0.95
#> itrf_E_2_APD  0.76
#> itrf_E_3_APD  0.58
#> itrf_E_4_APD  0.58
#> itrf_E_5_APD  0.93
#> itrf_E_6_APD  0.53
#> itrf_E_7_OPP  0.35
#> itrf_E_8_OPP  0.49
#> itrf_E_9_OPP  0.83
#> itrf_E_10_OPP 0.40
#> itrf_E_11_OPP 0.77
#> itrf_E_12_OPP 0.46
#> itrf_E_13_OPP 0.47
#> itrf_E_14_APD 0.38
#> itrf_E_15_APD 0.70
#> itrf_E_16_OPP 0.35