Scale Dictionary

A collection of procedures to bring labels, scales, missing values etc. to a data frame.

Examples

# apply a dictionary file to a data frame
dat <- apply_dic(ITRF, dic_ITRF)
# check for typos (not allowed values)
dat <- check_values(dat, replace = NA)
# Single imputation (EM algorith from the Amelia package)
# based on the variables of the provided scale
dat <- impute_missing(dat, scale = "ITRF_Ext")

#> No missing data.

dat <- impute_missing(dat, scale = "ITRF_Int")

#> No missing data.

# Show a table with all scales and scale labels included in the data frame
list_scales(dat, labels = TRUE)

#> # A tibble: 8 x 6
#>   Scale Subscale Subscale_2 `Label scale`      `Label subscale` `Label subscale…
#>   <fct> <fct>    <fct>      <fct>              <fct>            <fct>           
#> 1 ITRF  Ext      A          Integrated teache… Externalizing    Aggressiveness  
#> 2 ITRF  Ext      MA         Integrated teache… Externalizing    Maladjusted     
#> 3 ITRF  Int      A          Integrated teache… Internalizing    Aggressiveness  
#> 4 ITRF  Int      NE         Integrated teache… Internalizing    Negative Emotio…
#> 5 ITRF  Int      SH         Integrated teache… Internalizing    Shyness         
#> 6 ITRF  Int      SR         Integrated teache… Internalizing    Social rejection
#> 7 ITRF  Int      SU         Integrated teache… Internalizing    Suffering       
#> 8 ITRF  Int      NA         Integrated teache… Internalizing    NA              

# Example with pipeline syntax. Would be much easier to use the "describe" function
# from the psch packages instead of summarise_all here.
dat %>%
  select(get_index(dat, scale = "ITRF", subscale = "Ext")) %>%
  names2item(short = TRUE, id = TRUE) %>%
  summarise_all(mean, na.rm = TRUE) %>%
  round(2) %>%
  t()

#> Error in dat %>% select(get_index(dat, scale = "ITRF", subscale = "Ext")) %>%     names2item(short = TRUE, id = TRUE) %>% summarise_all(mean,     na.rm = TRUE) %>% round(2) %>% t(): could not find function "%>%"