Score scale Calculates the scale scores.

Score scale Calculates the scale scores.

score_scale(
  data,
  filter,
  scale = NULL,
  subscale = NULL,
  subscale_2 = NULL,
  bind = FALSE,
  sum = FALSE,
  label = NULL,
  FUN = NULL,
  ...
)

Arguments

data
A data frame

filter
A logical expression for any dic attribute (e.g. scale == "ITRF" & subscale == "Int")

bind
If set TRUE, returns the complete data frame. If set false, returns a data frame with the new score variables for each scale.

sum
If TRUE, function mean(x, na.rm = TRUE) is applied for buiding the scores. If FALSE, function sum(x, na.rm = TRUE) is applied. When argument FUN is set, `sum` is ignored.

label
A character string with a label for the resulting score variable. Automatically generated if label is not set.

FUN
A function for caluclating the score (e.g., median)

...
Further arguments passed to the FUN function (e.g., na.rm = TRUE)

Value

A data frame